Converted internal model to UInt16, simplified Node

+++ Cue/CueDown.swift
@@ -12,7 +12,7 @@ public class CueDown {
 	
 	public init() {}
 	
-	public func parse(document buffer: [UInt8]) -> Node {
+	public func parse(document buffer: [UInt16]) -> Node {
 		let parser = CueParser()
 		
 		parser.process(buffer)
